Minor corrections.

+++ main.cpp
@@ -11,7 +11,7 @@ int main()
     string::strcat(p, "test");
 
     std::cout << "p = " << p << std::endl << "p len = " << string::strlen(p) << std::endl << std::endl;
-    for (int i = 0; i < string::strlen(p); i++)
+    for (size_t i = 0; i < string::strlen(p); i++)
     {
         char* t = p + i;
         std::cout << "p[" << i << "] = " << *t << std::endl;
